requirements.txt
-i https://pypi.org/simple
async-generator==1.10 ; python_version >= '3.5'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
cryptography==41.0.1 ; python_version >= '3.7'
exceptiongroup==1.1.1 ; python_version < '3.11'
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
msal==1.22.0
outcome==1.2.0 ; python_version >= '3.7'
pycparser==2.21
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
pysocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0 ; python_version >= '3.7'
selenium==4.10.0
sniffio==1.3.0 ; python_version >= '3.7'
sortedcontainers==2.4.0
trio==0.22.0 ; python_version >= '3.7'
trio-websocket==0.10.3 ; python_version >= '3.7'
urllib3[socks]==2.0.3 ; python_version >= '3.7'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
tqdm
